app/requirements.txt
Flask
Flask-PyMongo
pandas
Werkzeug
python-dotenv